+++ session/session_legacy_test.go
@@ -21,7 +21,6 @@ import (
 	"net"
 	"os"
 	"path"
-	"runtime"
 	"sort"
 	"strconv"
 	"strings"
@@ -30,7 +29,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/docker/go-units"
 	. "github.com/pingcap/check"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/failpoint"
@@ -93,7 +91,6 @@ var _ = SerialSuites(&testSessionSerialSuite{})
 var _ = SerialSuites(&testBackupRestoreSuite{})
 var _ = SerialSuites(&testTxnStateSerialSuite{})
 var _ = SerialSuites(&testStatisticsSuite{})
-var _ = SerialSuites(&testTiDBAsLibrary{})
 
 type testSessionSuiteBase struct {
 	cluster testutils.Cluster
@@ -128,7 +125,9 @@ type testStatisticsSuite struct {
 	testSessionSuiteBase
 }
 
-type testTiDBAsLibrary struct{}
+func TestT(t *testing.T) {
+	TestingT(t)
+}
 
 func clearStorage(store kv.Storage) error {
 	txn, err := store.Begin()
@@ -5891,33 +5890,6 @@ func (s *testSessionSuite) TestTemporaryTableInterceptor(c *C) {
 	val, err := snap.Get(context.Background(), k)
 	c.Assert(err, IsNil)
 	c.Assert(val, BytesEquals, []byte("v1"))
-}
-
-func (s *testTiDBAsLibrary) TestMemoryLeak(c *C) {
-	initAndCloseTiDB := func() {
-		store, err := mockstore.NewMockStore(mockstore.WithStoreType(mockstore.EmbedUnistore))
-		c.Assert(err, IsNil)
-		defer store.Close()
-
-		dom, err := session.BootstrapSession(store)
-		//nolint:staticcheck
-		defer dom.Close()
-		c.Assert(err, IsNil)
-	}
-
-	runtime.GC()
-	memStat := runtime.MemStats{}
-	runtime.ReadMemStats(&memStat)
-	oldHeapInUse := memStat.HeapInuse
-
-	for i := 0; i < 20; i++ {
-		initAndCloseTiDB()
-	}
-
-	runtime.GC()
-	runtime.ReadMemStats(&memStat)
-	// before the fix, initAndCloseTiDB for 20 times will cost 900 MB memory, so we test for a quite loose upper bound.
-	c.Assert(memStat.HeapInuse-oldHeapInUse, Less, uint64(300*units.MiB))
 }
 
 func (s *testSessionSuite) TestTiDBReadStaleness(c *C) {
